dataprocess/pyproject.toml
[tool.poetry]
name = "bridger_dataprocess"
version = "0.1.0"
description = "Data processing for Bridger"
authors = ["Jason Portenoy"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8,<4"
humanfriendly = "^10.0"
requests = "^2.28.1"
backoff = "^2.2"
python-dotenv = "^0.20.0"
pandas = "^1.4.3"
tqdm = "^4.64.0"
scispacy = "^0.5.0"
pyarrow = "^9.0.0"
poetry = "^1.1.15"
langid = "^1.1.6"
sentence-transformers = "^2.2.2"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
black = "^22.6.0"
beaker-py = "^1.18.4"
SQLAlchemy = "^1.4.41"
psycopg2-binary = "^2.9.3"
sqlalchemy-redshift = "^0.8.11"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
